+++ packages/core/src/index.ts
@@ -3,3 +3,4 @@ export * from "./errors.js";
 export * from "./ids.js";
 export * from "./models.js";
 export * from "./protocol.js";
+export * from "./settings.js";